moved contact endpoint functionality to its own type
